+++ src/maincli.cpp
@@ -2,6 +2,7 @@
 #include <iostream>
 #include <fstream>
 #include <filesystem>
+#include <algorithm>
 #include "parser/dbcscanner.h"
 #include "codegen/c-main-generator.h"
 #include "codegen/c-util-generator.h"
@@ -9,6 +10,9 @@
 #include "codegen/version.h"
 
 #define GEN_UTIL_CODE
+
+#define MIN_ARGC_NUM 4
+#define MAX_ARGC_NUM 5
 
 char verstr[128] = {0};
 
@@ -18,8 +22,10 @@ const char* helptext =
   "To use utility you need to provide 3 arguments:\n\n"
   "1. dbc file path\n"
   "2. directory for generated source files (existable)\n"
-  "3. prefix (driver name) which will be used for naming dirver parts\n\n"
-  "Usage example:\n\n./dbccoder /home/user/docs/driveshaft.dbc /home/user/docs/gen/ drivedb\n\n";
+  "3. prefix (driver name) which will be used for naming dirver parts\n"
+  "4. (optional) --node-utils will generate specific pairs of binutils drivers for each node"
+  "   which is either transmitter or receiver of at least one frame from dbc.\n\n"
+  "Usage example:\n\n./dbccoder /home/user/docs/driveshaft.dbc /home/user/docs/gen/ drivedb --node-utils\n\n";
 
 DbcScanner* scanner;
 CiMainGenerator* cigen;
@@ -40,7 +46,7 @@ int main(int argc, char* argv[])
   std::snprintf(verstr, 128, "\nDbccoder v%u.%u\n\n", CODEGEN_LIB_VERSION_MAJ, CODEGEN_LIB_VERSION_MIN);
   std::cout << verstr;
 
-  if (argc == 4)
+  if (argc >= MIN_ARGC_NUM)
   {
     std::ifstream reader;
     // copy dbc file name to string variable
@@ -79,18 +85,89 @@ int main(int argc, char* argv[])
 
 #if defined (GEN_UTIL_CODE)
 
-    ret = fscreator->PrepareDirectory(dbc_driver_name.c_str(), (source_files_out_path).c_str(), true, info);
+    std::string node_util = "--node-utils";
 
-    MsgsClassification groups;
-
-    for (size_t i = 0; i < scanner->dblist.msgs.size(); i++)
+    // check if option --node-utils is requested, when requested binutil generation
+    // wiil be performed on each node from DBC file in accordance to its RX / TX subscription
+    if (argc == MAX_ARGC_NUM && node_util.compare(argv[4]) == 0)
     {
-      groups.Rx.push_back(scanner->dblist.msgs[i]->MsgID);
+      std::vector<std::string> nodes;
+
+      for (size_t num = 0; num < scanner->dblist.msgs.size(); num++)
+      {
+        // iterate all messages and collect All nodes assign to at least one message
+        auto m = scanner->dblist.msgs[num];
+
+        // test transmitter
+        std::string nodename = m->Transmitter;
+        bool found = (std::find(nodes.begin(), nodes.end(), nodename) != nodes.end());
+
+        if (!found)
+        {
+          nodes.push_back(nodename);
+        }
+
+        for (size_t recs = 0; recs < m->RecS.size(); recs++)
+        {
+          // test all recs
+          found = (std::find(nodes.begin(), nodes.end(), nodename) != nodes.end());
+
+          if (!found)
+          {
+            nodes.push_back(nodename);
+          }
+        }
+      }
+
+      // for each node in collection perform specific bin-util generation
+      for (size_t node = 0; node < nodes.size(); node++)
+      {
+        std::string util_name = nodes[node] + "_" + dbc_driver_name;
+
+        ret = fscreator->PrepareDirectory(util_name.c_str(), source_files_out_path.c_str(), true, info);
+
+        MsgsClassification groups;
+
+        for (size_t i = 0; i < scanner->dblist.msgs.size(); i++)
+        {
+          auto m = scanner->dblist.msgs[i];
+
+          if (m->Transmitter.compare(nodes[node]) == 0)
+          {
+            groups.Tx.push_back(m->MsgID);
+          }
+          else
+          {
+            bool found = (std::find(m->RecS.begin(), m->RecS.end(), nodes[node]) != m->RecS.end());
+
+            if (found)
+            {
+              groups.Rx.push_back(m->MsgID);
+            }
+          }
+        }
+
+        if (ret)
+        {
+          ciugen->Generate(scanner->dblist, fscreator->FS, groups, dbc_driver_name);
+        }
+      }
     }
-
-    if (ret)
+    else
     {
-      ciugen->Generate(scanner->dblist, fscreator->FS, groups, dbc_driver_name);
+      ret = fscreator->PrepareDirectory(dbc_driver_name.c_str(), (source_files_out_path).c_str(), true, info);
+
+      MsgsClassification groups;
+
+      for (size_t i = 0; i < scanner->dblist.msgs.size(); i++)
+      {
+        groups.Rx.push_back(scanner->dblist.msgs[i]->MsgID);
+      }
+
+      if (ret)
+      {
+        ciugen->Generate(scanner->dblist, fscreator->FS, groups, dbc_driver_name);
+      }
     }
 
 #endif